Renames link component.
