+++ packages/inertia-svelte/src/index.js
@@ -1,5 +1,5 @@
 export { default as App, default as app, default as InertiaApp } from './App.svelte'
-export { default as Link, default as link, default as InertiaLink } from './Link.svelte'
-export { default as inertia } from './link.js'
-export { default as page } from './page.js'
-export { default as remember } from './remember.js'
+export { default as Link, default as link, default as InertiaLink } from './InertiaLink.svelte'
+export { default as inertia } from './link'
+export { default as page } from './page'
+export { default as remember } from './remember'
